Исправлен баг работы с Agilent 3458a и Акип В7-78/1
git-svn-id: https://5-10/svn/mxsrclib/trunk@478 63356120-0428-6b47-bcd6-cd59724df219

+++ meassup.cpp
@@ -243,12 +243,12 @@ irs_bool u309m_current_supply_t::dc_supported()
 // Установка тока
 void u309m_current_supply_t::set_current(double a_current)
 {
-  m_current = a_current;
+  m_current = static_cast<float>(a_current);
 }
 // Установка напряжения
 void u309m_current_supply_t::set_voltage(double a_voltage)
 {
-  m_voltage = a_voltage;
+  m_voltage = static_cast<float>(a_voltage);
 }
 // Включение источника
 void u309m_current_supply_t::on()
